feat: Modified data type of NodeQuarantined from bool to Status
#### Customer Summary (required for feat, fix, and perf MRs)
A node is quarantined or not is identified by boolean value. But this is not sufficient since we don't want to remediate a node if it's already quarantined. To handle this use case replace boolean with Status data type

#### Testing
Tested this change with Fault remediation module. After added node states node is going into remediation only if node is Quarantined and podevictionstatus is succeeded  

#### Tickets: (use one of the action keywords: Closes / Fixes / Resolves / Relates to; NO-REF for issues without a tracking issue)

- JIRA: Closes NGCC-23689

Closes NGCC-23689

See merge request dgxcloud/mk8s/k8s-addons/nvsentinel!165