chore: remove logLevel from config
#### Customer Summary (required for feat, fix, and perf MRs)

logLevel was not getting use anywhere. We were validating it but it was not referenced in the code.
We are already setting the verbosity in the deployment file of csp-health-monitor and that verbosity is configurable from values.yaml. The same pattern we are following for other modules also.


Closes NGCC-26051

Closes NGCC-26051

See merge request dgxcloud/mk8s/k8s-addons/nvsentinel!258