feat: add count query to get documents inserted after an _id
#### TL;DR

The `ChangeStreamWatcher` is updated to support monitoring the number of unprocessed events. A new method, `GetUnprocessedEventCount`, is added to efficiently count new documents in the collection since the last event was processed. To support this, the watcher now stores the database and collection names. Additionally, the MongoDB read preference is set to `Primary` for token and client collections to ensure stronger consistency.

NOTE: There will be a follow up MR after this to use the updated version of the client in other components

#### Tickets:

- JIRA: Closes [\[KACE-307\] Add a metric to measure processing lag in fault quarantine module - JIRA SW](https://jirasw.nvidia.com/browse/KACE-307)

<!--Add a changelog entry for your MR so when we cut a tag, a changelog will be produced 
Changelog: <added/changed/removed>-->

<!--This is an auto-generated comment: release notes by coderabbit.ai-->

## Summary by CodeRabbit

* **New Features**
  * Added the ability to view the count of new unprocessed events since the last check, with support for additional filtering.
* **Bug Fixes**
  * Improved data consistency by enforcing a primary read preference for critical operations.

<!--end of auto-generated comment: release notes by coderabbit.ai-->

See merge request dgxcloud/mk8s/k8s-addons/nvsentinel!345