chore: ensure argocd reruns job instead of failing due to immutable spec
#### Customer Summary (required for feat, fix, and perf MRs)

If any spec in the database creation job is changed, the application has to manually force synced. This is not ideal as we want to get to a point where we can make MRs and ArgoCD does the right thing. 

As per the ArgoCD [docs](https://argo-cd.readthedocs.io/en/stable/user-guide/sync-options/#force-sync), the `Force=true,Replace=true` annotation can be used for:

```
For certain resources you might want to delete and recreate, e.g. job resources that should run every time when syncing.
```

This fits our use case since we want to rerun the job on every install to make sure any new changes to the job spec are honoured. The job itself is idempotent so there shouldn't be a problem rerunning it. 

#### TL;DR

- Added `Force=true,Replace=true`

#### Additional Details (optional for docs, build, test, refactor, ci, chore, style, and revert MRs)

N/A

#### For the Reviewer

N/A

#### For QA (optional for docs, build, test, refactor, ci, chore, style, and revert MRs)

N/A, following guidance in ArgoCD docs

#### Tickets: (use one of the action keywords: Closes / Fixes / Resolves / Relates to; NO-REF for issues without a tracking issue)

- JIRA: Closes NO-REF

<!--  Add a changelog entry for your MR so when we cut a tag, a changelog will be produced 
Changelog: <added/changed/removed>
-->

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

* **Chores**
  * Updated job configuration to improve synchronization behavior with Argo CD.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

See merge request dgxcloud/mk8s/k8s-addons/nvsentinel!365